pyproject.toml
[project]
name = "repl-nix-smartsupportai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.40.2",
]
